Added some casts to ensure 64bit arithmetic is done correctly on 32bit machines.
git-svn-id: http://svn.bramp.net:8080/svn/PhD/Work/userspace/threadnetperf@2733 e02701c3-087c-644a-a2a3-551e05f81084

+++ common.c
@@ -178,7 +178,7 @@ unsigned long long get_nanoseconds() {
 
 	//getnstimeofday(&ts, NULL);
 	clock_gettime(CLOCK_REALTIME, &ts);
-	nanoseconds = ts.tv_sec * 1000000000 + ts.tv_nsec;
+	nanoseconds = (unsigned long long)ts.tv_sec * 1000000000 + (unsigned long long)ts.tv_nsec;
 #endif
 
 	return nanoseconds;
